- Pembuatan fitur untuk melihat komentar. - Pembuatan fitur untuk menambahkan komentar. - Refactor kodingan cara pengambilan data di halaman detail post. - Penyesuaian fitur penghapusan post dengan menambahkan dialog konfirmasi. - Penyesuaian fitur logout dengan menambahkan dialog konfirmasi. - Penyesuaian fitur tambah postingan untuk menambahkan cabang kosong untuk like & comment.

+++ app/src/main/java/com/example/montessori/ui/user/PostDetailActivity.java
@@ -13,13 +13,18 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 import com.example.montessori.R;
+import com.example.montessori.adapter.CommentAdapter;
+import com.example.montessori.model.Comment;
 import com.example.montessori.model.PostMember;
 import com.example.montessori.ui.ImagePreviewActivity;
 import com.example.montessori.util.Helper;
@@ -45,13 +50,17 @@ public class PostDetailActivity extends AppCompatActivity {
     private final FirebaseFirestore database = FirebaseFirestore.getInstance();
     private final CollectionReference postReference = database.collection(ReferenceConstant.ALL_POSTS);
     private final CollectionReference likeReference = database.collection(ReferenceConstant.LIKE);
+    private final CollectionReference commentsReference = database.collection(ReferenceConstant.COMMENTS);
 
     private TaskCompletionSource<PostMember> postSource = new TaskCompletionSource<>();
     private TaskCompletionSource<List<String>> likeSource = new TaskCompletionSource<>();
+    private final TaskCompletionSource<List<Comment>> commentSource = new TaskCompletionSource<>();
 
-    // private TaskCompletionSource<PostMember> commentSource = new TaskCompletionSource<>();
+    private CommentAdapter commentAdapter;
 
     private PostMember postData;
+    private HashMap<String, Object> likeData = new HashMap<>();
+    private final ArrayList<Comment> commentData = new ArrayList<>();
     // NOTE: Untuk simpan state mengenai apakah postingan sudah dilike atau belum oleh pengguna.
     private boolean isPostLiked = false;
 
@@ -61,6 +70,7 @@ public class PostDetailActivity extends AppCompatActivity {
     private ProgressBar progressBar;
     private ImageView ivPost, ivLike;
     private TextView tvTitle, tvName, tvCategory, tvLikeCount, tvDesc;
+    private RecyclerView rvComment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,6 +78,7 @@ public class PostDetailActivity extends AppCompatActivity {
         setContentView(R.layout.activity_post_detail);
 
         postData = getIntent().getParcelableExtra(IntentNameExtra.POST_DATA);
+        commentAdapter = new CommentAdapter(this);
 
         if (postData == null) {
             Toast.makeText(this, "Data tidak ditemukan.", Toast.LENGTH_SHORT).show();
@@ -83,6 +94,8 @@ public class PostDetailActivity extends AppCompatActivity {
         commentContainer = findViewById(R.id.commentContainer);
         progressBar = findViewById(R.id.progressBar);
 
+        rvComment = findViewById(R.id.rvComment);
+
         ivPost = findViewById(R.id.ivPost);
         tvTitle = findViewById(R.id.tvTitle);
         tvName = findViewById(R.id.tvName);
@@ -92,27 +105,46 @@ public class PostDetailActivity extends AppCompatActivity {
 
         ImageButton btnBack = findViewById(R.id.btnBack);
         ImageButton btnDelete = findViewById(R.id.btnDelete);
+
         ivLike = findViewById(R.id.btnLike);
         Button btnShare = findViewById(R.id.btnShare);
 
         swipeLayout.setOnRefreshListener(this::loadData);
         statusContainer.setOnClickListener(view -> Toast.makeText(this, "The post hasn't approved by admin yet so the like and comment feature is still disabled.", Toast.LENGTH_LONG).show());
 
+        rvComment.setLayoutManager(new LinearLayoutManager(this));
+
         btnBack.setOnClickListener(view -> finish());
         btnDelete.setVisibility(currentUser.getUid().equals(postData.getUid()) ? View.VISIBLE : View.GONE);
-        // Harusnya ada pop-up konfirmasi
-        // AlertDialog confirmation
-        btnDelete.setOnClickListener(view -> {
+        btnDelete.setOnClickListener(this::showAlertDialog);
+        btnShare.setOnClickListener(view -> startActivity(new Intent(PostDetailActivity.this, ShareActivity.class)));
+
+        likeContainer.setOnClickListener(view -> doLikePost());
+        commentContainer.setOnClickListener(view -> {
+            Intent commentIntent = new Intent(this, CommentsActivity.class);
+            commentIntent.putExtra(IntentNameExtra.POST_ID_DATA, postData.getId());
+            startActivity(commentIntent);
+        });
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        loadData();
+    }
+
+    private void showAlertDialog(View v) {
+        AlertDialog.Builder alert = new AlertDialog.Builder(this);
+        alert.setMessage("Do you want to delete this post?");
+        alert.setPositiveButton("Yes", ((dialogInterface, i) -> {
             if (currentUser.getUid().equals(postData.getUid())) {
                 deleteData(postData.getId());
                 finish();
             }
+        }));
+        alert.setNegativeButton("No", (dialogInterface, i) -> {
         });
-
-        likeContainer.setOnClickListener(view -> doLikePost());
-        btnShare.setOnClickListener(view -> startActivity(new Intent(PostDetailActivity.this, ShareActivity.class)));
-
-        loadData();
+        alert.create().show();
     }
 
     private void doLikePost() {
@@ -126,7 +158,7 @@ public class PostDetailActivity extends AppCompatActivity {
                     toggleLikeButton();
                     loadLikes(true);
                 } else {
-                    Toast.makeText(this, "Terjadi kesalahan.", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(this, "Error occurred when disliking the post.", Toast.LENGTH_SHORT).show();
                 }
             });
         } else {
@@ -138,7 +170,7 @@ public class PostDetailActivity extends AppCompatActivity {
                     toggleLikeButton();
                     loadLikes(true);
                 } else {
-                    Toast.makeText(this, "Terjadi kesalahan.", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(this, "Error occurred when liking the post.", Toast.LENGTH_SHORT).show();
                 }
             });
         }
@@ -146,7 +178,7 @@ public class PostDetailActivity extends AppCompatActivity {
 
     private void setPostData(PostMember postData) {
         if (postData == null) {
-            Toast.makeText(this, "Error occurred when load post data", Toast.LENGTH_SHORT).show();
+            Toast.makeText(this, "Error occurred when load post data.", Toast.LENGTH_SHORT).show();
             finish();
             return;
         }
@@ -174,32 +206,37 @@ public class PostDetailActivity extends AppCompatActivity {
     private void deleteData(String postId) {
         Task<Void> postTask = postReference.document(postId).delete();
         Task<Void> likeTask = likeReference.document(postId).delete();
+        Task<Void> commentTask = commentsReference.document(postId).delete();
 
-        Tasks.whenAll(postTask, likeTask).addOnCompleteListener(task -> {
-            if (!task.isSuccessful()) {
-                Toast.makeText(PostDetailActivity.this, "Error occurred when deleting the post.", Toast.LENGTH_SHORT).show();
-            } else {
+        Tasks.whenAll(postTask, likeTask, commentTask).addOnCompleteListener(task -> {
+            // NOTE: Kalau gagal menghapus detail postingan namun data like sudah terhapus kembalikan data like yg terakhir diambil.
+            if (!postTask.isSuccessful()) {
+                if (likeTask.isSuccessful()) {
+                    likeReference.document(postId).set(likeData);
+                }
+                if (commentTask.isSuccessful()) {
+                    commentsReference.document(postId).set(commentData);
+                }
+            }
+
+            if (task.isSuccessful()) {
                 Toast.makeText(PostDetailActivity.this, "Post deleted successfully", Toast.LENGTH_SHORT).show();
                 finish();
+            } else {
+                Toast.makeText(PostDetailActivity.this, "Error occurred when deleting the post.", Toast.LENGTH_SHORT).show();
             }
         });
     }
 
     private void loadData() {
         toggleLoadingVisibility(true);
-        // NOTE: Ketika post belum dicek oleh admin atau tidak diterima, ambil data postingan saja.
-        if (postData.isNotCheckedOrApproved()) {
-            loadPostDetail(true);
-            return;
-        }
-
         init();
 
-        Tasks.whenAll(postSource.getTask(), likeSource.getTask()).addOnCompleteListener(task -> {
+        Tasks.whenAll(postSource.getTask(), likeSource.getTask(), commentSource.getTask()).addOnCompleteListener(task -> {
             if (postSource.getTask().isSuccessful() && postSource.getTask().getResult() != null) {
                 setPostData(postSource.getTask().getResult());
             } else {
-                Toast.makeText(this, "Error occurred when load post data", Toast.LENGTH_SHORT).show();
+                setPostData(null);
             }
 
             if (likeSource.getTask().isSuccessful() && likeSource.getTask().getResult() != null) {
@@ -213,36 +250,35 @@ public class PostDetailActivity extends AppCompatActivity {
                 }
                 tvLikeCount.setText(Helper.getTotalLikes(PostDetailActivity.this, userLikes.size()));
             } else {
-                Toast.makeText(this, "Error occurred when load like data", Toast.LENGTH_SHORT).show();
+                Toast.makeText(this, "Error occurred when load like data.", Toast.LENGTH_SHORT).show();
+            }
+
+            if (commentSource.getTask().isSuccessful() && commentSource.getTask().getResult() != null) {
+                List<Comment> commentsList = commentSource.getTask().getResult();
+                commentAdapter.setLastComment(commentsList);
+                rvComment.setAdapter(commentAdapter);
+                rvComment.setVisibility(!commentsList.isEmpty() ? View.VISIBLE : View.GONE);
+            } else {
+                Toast.makeText(this, "Error occurred when load comment data.", Toast.LENGTH_SHORT).show();
+                rvComment.setVisibility(View.GONE);
             }
 
             toggleLoadingVisibility(false);
         });
 
-
-        loadPostDetail(false);
+        loadPostDetail();
         loadLikes(false);
-        // loadComments();
+        loadComments(false);
     }
 
-    // NOTE: isSingleLoading di loadPostDetail untuk membedakan postingan sudah diapprove/belum sehingga tidak perlu mengambil semua data
-    // Ada kemungkinan nanti dihapus
-    private void loadPostDetail(boolean isSingleLoading) {
+    private void loadPostDetail() {
         postReference.document(postData.getId()).get().addOnCompleteListener(task -> {
             toggleLoadingVisibility(false);
             if (task.getResult() != null && task.getResult().exists()) {
                 PostMember data = task.getResult().toObject(PostMember.class);
-                if (!isSingleLoading) {
-                    postSource.trySetResult(data);
-                } else {
-                    setPostData(data);
-                }
+                postSource.trySetResult(data);
             } else {
-                if (!isSingleLoading) {
-                    postSource.trySetException(new Exception("Error occurred"));
-                } else {
-                    setPostData(null);
-                }
+                postSource.trySetException(new Exception("Error occurred"));
             }
         });
     }
@@ -252,6 +288,7 @@ public class PostDetailActivity extends AppCompatActivity {
         likeReference.document(postData.getId()).get().addOnCompleteListener(task -> {
             if (task.isSuccessful() && task.getResult() != null && task.getResult().exists() && task.getResult().getData() != null) {
                 Map<String, Object> likes = task.getResult().getData();
+                likeData = new HashMap<>(likes);
                 ArrayList<String> userLikes = new ArrayList<>();
                 for (Map.Entry<String, Object> user : likes.entrySet()) {
                     userLikes.add(user.getKey());
@@ -271,8 +308,40 @@ public class PostDetailActivity extends AppCompatActivity {
         });
     }
 
-    private void loadComments() {
+    private void loadComments(boolean isSingleLoading) {
+        commentsReference.document(postData.getId()).get().addOnCompleteListener(task -> {
+            if (task.isSuccessful() && task.getResult() != null && task.getResult().exists() && task.getResult().getData() != null) {
+                Map<String, Object> comments = task.getResult().getData();
+                ArrayList<Comment> commentsList = new ArrayList<>();
+                for (Map.Entry<String, Object> comment : comments.entrySet()) {
+                    Comment data = new Comment();
+                    HashMap<String, Object> commentData = (HashMap<String, Object>) comment.getValue();
+                    data.setComment((String) commentData.get(Comment.COMMENT_FIELD));
+                    data.setName((String) commentData.get(Comment.NAME_FIELD));
+                    data.setDate((String) commentData.get(Comment.DATE_FIELD));
+                    data.setTimestamp((Long) commentData.get(Comment.TIMESTAMP_FIELD));
+                    commentsList.add(data);
+                }
 
+                // Simpan backup data.
+                this.commentData.clear();
+                this.commentData.addAll(commentsList);
+                // Kirim data yang didapat.
+                commentSource.trySetResult(commentsList);
+                /*if (!isSingleLoading) {
+                    commentSource.trySetResult(commentsList);
+                } else {
+                    tvLikeCount.setText(Helper.getTotalLikes(PostDetailActivity.this, userLikes.size()));
+                }*/
+            } else {
+                commentSource.trySetException(new Exception("Get comment fail"));
+                /*if (!isSingleLoading) {
+                    commentSource.trySetException(new Exception("Get like fail"));
+                } else {
+                    Toast.makeText(this, "Error occurred when load like data", Toast.LENGTH_SHORT).show();
+                }*/
+            }
+        });
     }
 
     private void init() {
